refactor: move message interface into network package
