+++ dsl/src/commonMain/kotlin/it/unibo/collektive/Collektive.kt
@@ -2,7 +2,7 @@ package it.unibo.collektive
 
 import it.unibo.collektive.aggregate.AggregateContext
 import it.unibo.collektive.aggregate.AggregateResult
-import it.unibo.collektive.messages.InboundMessage
+import it.unibo.collektive.networking.InboundMessage
 import it.unibo.collektive.networking.Network
 import it.unibo.collektive.state.State
