+++ dsl/src/commonTest/kotlin/it/unibo/collektive/network/NetworkImplTest.kt
@@ -1,8 +1,8 @@
 package it.unibo.collektive.network
 
 import it.unibo.collektive.ID
-import it.unibo.collektive.messages.InboundMessage
-import it.unibo.collektive.messages.OutboundMessage
+import it.unibo.collektive.networking.InboundMessage
+import it.unibo.collektive.networking.OutboundMessage
 import it.unibo.collektive.networking.Network
 
 class NetworkImplTest(private val networkManager: NetworkManager, private val localId: ID) : Network {
